switched threading to multiprocessing.
processes are stored in a list called processes.
p1 takes the place of t1, etc.

Something is still hung up at the end of the processes. Perphaps it has something to do with the process that is not active/being utilized by the client? Perhaps killing that process will end the deadlock.